# Conflicts:
#	src/main/java/pro/sky/telegrambotshelter/listener/TelegramBotUpdatesListener.java

+++ src/main/java/pro/sky/telegrambotshelter/service/ShelterService.java
@@ -15,7 +15,7 @@ public class ShelterService {
     }
 
     /**метод для получения id волонтера*/
-    public int getVolunteerChatId(String shelterType) {
+    public long getVolunteerChatId(String shelterType) {
         return shelterRepository.findShelterByShelterType(shelterType).getVolunteerChatId();
     }
